Start to break up the monolithic `app.py` and add types
